move svelte-adapter into packages
