+++ packages/adapter-svelte/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "../../svelte"
+	},
+	"include": ["./src/**/*.ts"]
+}
