--- tsconfig-svelte.json
@@ -1,8 +1,0 @@
-{
-	"extends": "./tsconfig.json",
-	"compilerOptions": {
-		"module": "ES2020",
-		"outDir": "./svelte"
-	},
-	"include": ["./src/svelte-store.ts"]
-}
